Added VideoPlayer and VideLikes component

+++ eslint.config.js
@@ -21,7 +21,7 @@ export default defineConfig([
     },
     rules: {
       semi: ['error', 'always'],
-      quotes: ['error', 'single'],
+      quotes: ['error', 'double'],
       'no-unused-vars': 'warn',
       'react-hooks/rules-of-hooks': 'error',
       'react-hooks/exhaustive-deps': 'warn',
